added check for translation existence

+++ src/main/java/com/brigade22/spring/springlabs/services/DictionaryService.java
@@ -125,6 +125,10 @@ public class DictionaryService {
     }
 
     public void deleteTranslationById(int id) {
+        Translation translation = translationRepository.findById(id);
+        if (translation == null) {
+            throw new ResourceNotFoundException("Translation with id " + id + " not found");
+        }
         translationRepository.deleteTranslationById(id);
     }
 }
